Add rich native notifications
Introduce a full native notifications feature: simple and rich notifications, builder APIs, async activations and cancellation, and platform implementations.

Highlights:
- Public API: builder methods (EnableNotifications, SetDefaultNotificationIcon), feature interfaces/extensions, ShowNotification overloads and ShowNotificationAsync.
- Shared types: options, action, activation, result, urgency records/enums.
- Native bridge: new exports (ShowNotificationWithOptions, BeginShowNotification, CancelNotification) and parameter plumbing for DefaultNotificationIcon.
- Platform implementations for Windows (WinToast), macOS (UNUserNotificationCenter), and Linux (libnotify).
- Operation handling: InfiniNotificationOperation for async callbacks/cancellation.
- Tests, docs (notifications guide), and sidebar entry added.

+++ tests/InfiniAutomationTests/Tests/ZoomDpiParityAutomationTests.cs
@@ -3,7 +3,6 @@
 // ---------------------------------------------------------------------------------------------------------------------
 using InfiniTests;
 using Microsoft.Playwright;
-using System.Text.Json;
 
 namespace InfiniAutomationTests.Tests;
 // ---------------------------------------------------------------------------------------------------------------------
@@ -61,6 +60,6 @@ public abstract class ZoomDpiParityAutomationTests : InfiniFramePlaywrightTestBa
         );
 
         await Assert.That(nativeDpi).IsGreaterThan(0);
-        await Assert.That(jsDpi).IsEqualTo((double)nativeDpi);
+        await Assert.That(jsDpi).IsEqualTo(nativeDpi);
     }
 }
